Made corrections as per feedback provided
