+++ express_server.js
@@ -1,10 +1,12 @@
+const { getUserByEmail, urlsForUser, generateRandomString, userexist } = require("./helpers.js");
+const bodyParser = require("body-parser");
 const express = require("express");
 const app = express();
 const PORT = 8080; // default port 8080
 const cookieSession = require('cookie-session');
 const bcrypt = require('bcrypt');
 
-app.set("view engine", "ejs");
+//////SETTING UP DATABASES///////
 
 let users = {
   "userRandomID": {
@@ -30,37 +32,33 @@ let urlDatabase = {
   }
 };
 
-const bodyParser = require("body-parser");
-const getUserByEmail = require("./helpers");
+/////HELPER FUNCTIONS///////
+
+
+
+///////APP PREP///////
+app.set("view engine", "ejs");
+
 app.use(bodyParser.urlencoded({extended: true}));
 app.use(cookieSession({ name: 'session', keys: ['key1', 'key2'] }));
 
+///////APP GET ROUTES/////
+
 app.get("/", (req, res) => {
-  res.send("Hello!");
+  //res.send("Hello!");
+  res.redirect("/login");
 });
 
 app.get("/urls.json", (req, res) => {
   res.json(urlDatabase);
 });
 
-const urlsForUser = function(id) {
-  let filteredURLS = {};
-
-  for (let i in urlDatabase) {
-    if (urlDatabase[i].userID === id) {
-      filteredURLS[i] = urlDatabase[i].longURL;
-    }
-  }
-
-  return filteredURLS;
-};
-
 app.get("/urls", (req, res) => {
   const userid = req.session.user_id;
   const user = users[userid];
 
   if (user) {
-    const templateVars = { urls: urlsForUser(userid), user, userid };
+    const templateVars = { urls: urlsForUser(userid, users), user, userid };
     res.render("urls_index", templateVars);
   } else {
     res.redirect("/login");
@@ -94,6 +92,65 @@ app.get("/urls/new", (req, res) => {
   }
 
 });
+
+app.get("/login", (req,res) => {
+  const userid = req.session.user_id;
+
+  const user = users[userid];
+  console.log('user: ', user);
+  if (user) {
+    res.redirect('/urls');
+  } else {
+    const templateVars = { username: req.session ? req.session["username"] : "", user,userid};
+    res.render("urls_login", templateVars);
+  }
+ 
+});
+
+app.get("/urls/:shortURL", (req, res) => {
+  const userid = req.session.user_id;
+  const user = users[userid];
+  if (user) {
+    const urlsList = urlsForUser(userid, users);
+    if (urlsList[req.params.shortURL]) {
+      const shortURL = req.params.shortURL;
+      const templateVars = {
+        username: req.session ? req.session["username"] : "",
+        shortURL: shortURL,
+        longURL: urlDatabase[shortURL].longURL,
+        user,
+        userid
+      };
+      res.render("urls_show", templateVars);
+    } else {
+      res.redirect("/urls");
+    }
+  } else {
+    res.redirect("/login");
+  }
+});
+
+app.get("/u/:shortURL", (req, res) => {
+  if (urlDatabase[req.params.shortURL]) {
+    res.redirect(urlDatabase[req.params.shortURL].longURL);
+  } else {
+    res.status(404).send('Long URL not found!');
+  }
+});
+
+app.get("/register", (req, res) => {
+  const userid = req.session.user_id;
+  const user = users[userid];
+  if (user) {
+    res.redirect('/urls');
+  } else {
+    const templateVars = { username: req.session ? req.session["username"] : "", user,userid};
+    res.render("urls_register", templateVars);
+  }
+ 
+});
+
+///////APP POST ROUTES/////
 
 app.post("/urls", (req, res) => {
   const userid = req.session.user_id;
@@ -142,29 +199,30 @@ app.post("/register",(req,res) => {
 });
 
 app.post("/login",(req,res) => {
+  if(req.body.email === "" && req.body.password === ""){
+    res.status(403).send('Missing both email and password!');
+  }
 
-  const user = userexist(req.body.email, req.body.password);
+  if(req.body.email === ""){
+    res.status(403).send('Missing email');
+  }
+
+  if(req.body.password === ""){
+    res.status(403).send('Missing password');
+  }
+
+  const user = userexist(req.body.email, req.body.password, users);
 
   if (user) {
     req.session['user_id'] = user.id;
     res.redirect('/urls');
   } else {
-    res.status(403).send('User not found');
+    if(getUserByEmail(req.body.email, users)){
+      res.status(403).send('Wrong Password');
+    } else {
+      res.status(403).send('Email not found');
+    }
   }
-});
-
-app.get("/login", (req,res) => {
-  const userid = req.session.user_id;
-
-  const user = users[userid];
-  console.log('user: ', user);
-  if (user) {
-    res.redirect('/urls');
-  } else {
-    const templateVars = { username: req.session ? req.session["username"] : "", user,userid};
-    res.render("urls_login", templateVars);
-  }
- 
 });
 
 app.post("/logout",(req,res) => {
@@ -176,7 +234,7 @@ app.post("/urls/:shortURL/delete",(req,res) => {
   const userid = req.session.user_id;
   const user = users[userid];
   if (user) {
-    const urlsList = urlsForUser(userid);
+    const urlsList = urlsForUser(userid, users);
     if (urlsList[req.params.shortURL]) {
       delete urlDatabase[req.params.shortURL];
     }
@@ -187,88 +245,8 @@ app.post("/urls/:shortURL/delete",(req,res) => {
   }
 });
 
-app.get("/urls/:shortURL", (req, res) => {
-  const userid = req.session.user_id;
-  const user = users[userid];
-  if (user) {
-    const urlsList = urlsForUser(userid);
-    if (urlsList[req.params.shortURL]) {
-      const shortURL = req.params.shortURL;
-      const templateVars = {
-        username: req.session ? req.session["username"] : "",
-        shortURL: shortURL,
-        longURL: urlDatabase[shortURL].longURL,
-        user,
-        userid
-      };
-      res.render("urls_show", templateVars);
-    } else {
-      res.redirect("/urls");
-    }
-  } else {
-    res.redirect("/login");
-  }
-});
-
-app.get("/u/:shortURL", (req, res) => {
-  if (urlDatabase[req.params.shortURL]) {
-    res.redirect(urlDatabase[req.params.shortURL].longURL);
-  } else {
-    res.status(404).send('Long URL not found!');
-  }
-});
+///////APP PORT LISTENING FUNCTIONS////////
 
 app.listen(PORT, () => {
   console.log(`Example app listening on port ${PORT}!`);
 });
-
-app.get("/register", (req, res) => {
-  const userid = req.session.user_id;
-  const user = users[userid];
-  if (user) {
-    res.redirect('/urls');
-  } else {
-    const templateVars = { username: req.session ? req.session["username"] : "", user,userid};
-    res.render("urls_register", templateVars);
-  }
- 
-});
-
-const generateRandomInteger = function(min,max) {
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-const generateRandomString = function() {
-  let result = "";
-  
-  for (let i = 0; i < 6; i++) {
-    let choice = generateRandomInteger(1,3);
-    let randNum = 0;
-    if (choice === 1) {
-      randNum = generateRandomInteger(97,122);
-    } else if (choice === 2) {
-      randNum = generateRandomInteger(65,90);
-    } else {
-      randNum = generateRandomInteger(48,57);
-    }
-
-    let character = String.fromCharCode(randNum);
-    result += character;
-  }
-  console.log(result);
-  return result;
-};
-
-const userexist = function(inputemail,inputpassword) {
-  let result = null;
-  for (let user in users) {
-    let currentuser = users[user];// storing users index value into user
-    let storedemail = currentuser.email;
-
-    if (inputemail === storedemail && bcrypt.compareSync(inputpassword, currentuser.password)) {
-      return currentuser;
-    }
-  }
-
-  return result;
-};
