+++ views/urls_login.ejs
@@ -18,12 +18,12 @@
       <form action = "/login" method = "post">
         <div class="form-group">
           <label for="email">Email address</label>
-          <input id="email" type="email" class="form-control w-50" placeholder="name@example.com" name = 'email'>
+          <input id="email" type="email" class="form-control w-50" placeholder="Enter email..." name = 'email'>
 
         </div>
         <div class="form-group">
           <label for="pw">Password</label>
-          <input id="pw" type="Password" class="form-control w-50" placeholder="1234" name = 'password'>
+          <input id="pw" type="Password" class="form-control w-50" placeholder="Enter password..." name = 'password'>
         </div>
         <button type="submit" class="btn btn-primary mb-2">Login</button>
       </form>
